+++ api/src/org/labkey/api/reports/report/r/RserveScriptEngineFactory.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c) 2012 LabKey Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.labkey.api.reports.report.r;
+
+import org.labkey.api.reports.ExternalScriptEngineDefinition;
+import org.labkey.api.reports.ExternalScriptEngineFactory;
+
+import javax.script.ScriptEngine;
+
+public class RserveScriptEngineFactory extends ExternalScriptEngineFactory
+{
+    public RserveScriptEngineFactory(ExternalScriptEngineDefinition def)
+    {
+        super(def);
+    }
+
+    @Override
+    public synchronized ScriptEngine getScriptEngine()
+    {
+        return new RserveScriptEngine(_def);
+    }
+}
