+++ api/src/org/labkey/api/reports/report/r/view/HrefOutput.java
@@ -17,7 +17,7 @@
 package org.labkey.api.reports.report.r.view;
 
 import org.labkey.api.reports.Report;
-import org.labkey.api.reports.report.RReportDescriptor;
+import org.labkey.api.reports.report.r.RReportDescriptor;
 import org.labkey.api.reports.report.ScriptReportDescriptor;
 import org.labkey.api.view.HttpView;
 import org.labkey.api.view.ViewContext;
