+++ api/src/org/labkey/api/reports/report/r/view/KnitrOutput.java
@@ -17,7 +17,7 @@
 package org.labkey.api.reports.report.r.view;
 
 import org.jetbrains.annotations.Nullable;
-import org.labkey.api.reports.report.RReport;
+import org.labkey.api.reports.report.r.RReport;
 import org.labkey.api.reports.report.ScriptOutput;
 import org.labkey.api.reports.report.r.ParamReplacementSvc;
 import org.labkey.api.settings.AppProps;
