+++ api/src/org/labkey/api/reports/report/view/RenderBackgroundRReportView.java
@@ -16,7 +16,7 @@
 
 package org.labkey.api.reports.report.view;
 
-import org.labkey.api.reports.report.RReport;
+import org.labkey.api.reports.report.r.RReport;
 import org.labkey.api.reports.report.r.ParamReplacement;
 import org.labkey.api.reports.report.r.ParamReplacementSvc;
 import org.labkey.api.view.HttpView;
