+++ core/src/org/labkey/core/reports/ScriptEngineManagerImpl.java
@@ -47,10 +47,10 @@ import org.labkey.api.query.FieldKey;
 import org.labkey.api.reports.ExternalScriptEngineDefinition;
 import org.labkey.api.reports.ExternalScriptEngineFactory;
 import org.labkey.api.reports.LabKeyScriptEngineManager;
-import org.labkey.api.reports.RDockerScriptEngineFactory;
-import org.labkey.api.reports.RScriptEngineFactory;
+import org.labkey.api.reports.report.r.RDockerScriptEngineFactory;
+import org.labkey.api.reports.report.r.RScriptEngineFactory;
 import org.labkey.api.reports.RemoteRNotEnabledException;
-import org.labkey.api.reports.RserveScriptEngineFactory;
+import org.labkey.api.reports.report.r.RserveScriptEngineFactory;
 import org.labkey.api.script.RhinoScriptEngine;
 import org.labkey.api.script.ScriptService;
 import org.labkey.api.security.Encryption;
