+++ core/src/org/labkey/core/view/configReportsAndScripts.jsp
@@ -22,7 +22,7 @@
 <%@ page import="org.labkey.api.reports.ExternalScriptEngine" %>
 <%@ page import="org.labkey.api.reports.ExternalScriptEngineDefinition" %>
 <%@ page import="org.labkey.api.reports.report.ExternalScriptEngineReport" %>
-<%@ page import="org.labkey.api.reports.report.RReport" %>
+<%@ page import="org.labkey.api.reports.report.r.RReport" %>
 <%@ page import="org.labkey.api.reports.report.ScriptEngineReport" %>
 <%@ page import="org.labkey.api.rstudio.RStudioService" %>
 <%@ page import="org.labkey.api.security.permissions.AdminOperationsPermission" %>
