+++ pipeline/src/org/labkey/pipeline/api/ScriptTaskImpl.java
@@ -35,9 +35,8 @@ import org.labkey.api.reports.ExternalScriptEngine;
 import org.labkey.api.reports.ExternalScriptEngineDefinition;
 import org.labkey.api.reports.ExternalScriptEngineFactory;
 import org.labkey.api.reports.LabKeyScriptEngineManager;
-import org.labkey.api.reports.RScriptEngine;
-import org.labkey.api.reports.RserveScriptEngine;
-import org.labkey.api.services.ServiceRegistry;
+import org.labkey.api.reports.report.r.RScriptEngine;
+import org.labkey.api.reports.report.r.RserveScriptEngine;
 import org.labkey.api.settings.AppProps;
 import org.labkey.api.util.LogPrintWriter;
 import org.labkey.api.util.PageFlowUtil;
