+++ query/src/org/labkey/query/reports/ReportsController.java
@@ -76,7 +76,7 @@ import org.labkey.api.reports.RemoteRNotEnabledException;
 import org.labkey.api.reports.Report;
 import org.labkey.api.reports.ReportContentEmailManager;
 import org.labkey.api.reports.ReportService;
-import org.labkey.api.reports.RserveScriptEngine;
+import org.labkey.api.reports.report.r.RserveScriptEngine;
 import org.labkey.api.reports.actions.ReportForm;
 import org.labkey.api.reports.model.DataViewEditForm;
 import org.labkey.api.reports.model.ViewCategory;
@@ -87,8 +87,8 @@ import org.labkey.api.reports.report.AbstractReport;
 import org.labkey.api.reports.report.AbstractReportIdentifier;
 import org.labkey.api.reports.report.ModuleReportIdentifier;
 import org.labkey.api.reports.report.QueryReport;
-import org.labkey.api.reports.report.RReport;
-import org.labkey.api.reports.report.RReportJob;
+import org.labkey.api.reports.report.r.RReport;
+import org.labkey.api.reports.report.r.RReportJob;
 import org.labkey.api.reports.report.ReportDescriptor;
 import org.labkey.api.reports.report.ReportIdentifier;
 import org.labkey.api.reports.report.ReportUrls;
